+++ src/main/java/com/aleyla/footballTeam/service/PlayerService.java
@@ -24,11 +24,15 @@ public class PlayerService {
         return player;
     }
     public Player save(Player player){
-        //TODO Controller
+        validatePlayer(player);
         return playerRepository.save(player);
     }
 
     public void delete(Long id){
+        Player player = playerRepository.findById(id).orElse(null);
+        if (player == null) {
+            throw new EntityNotFoundException();
+        }
         playerRepository.deleteById(id);
     }
 
@@ -39,9 +43,24 @@ public class PlayerService {
         if (player.getId() == null) {
             player.setId(id);
         }
-        //TODO Controller
-
+        validatePlayer(player);
         return playerRepository.save(player);
+    }
+
+
+    private void validatePlayer(Player player) {
+        if (player == null) {
+            throw new InvalidRequestException("player", player, "PLAYER_COULD_NOT_BE_EMPTY");
+        }
+        if (player.getName() == null || player.getName().isEmpty()) {
+            throw new InvalidRequestException("name", player.getName(), "TEAM_NAME_COULD_NOT_BE_EMPTY");
+        }
+        if (player.getSurname() == null || player.getSurname().isEmpty()) {
+            throw new InvalidRequestException("surname",  player.getSurname(), "PLAYER_SURNAME_COULD_NOT_BE_EMPTY");
+        }
+        if (player.getBirthday() == null) {
+            throw new InvalidRequestException("birthday",  player.getBirthday(), "PLAYER_SURNAME_COULD_NOT_BE_EMPTY");
+        }
     }
 
     public List<Player> findAll() {
